Pulse CONVST before read/write commands
Realized every read/write command must be preceded by a pulse to CONVST pin.
The consequence of that is that reading/writing from/to AD7091R8
requires a GPIO to pulse the CONVST line.
That was not needed for AD7091R nor AD7091R-5 devices because:
- The CONVST in not tied to SDI as it is in AD7091R.
- In command mode, the AD7091R-5 converts the next programmed channel
  when the conversion result register is read.
Thus, require gpio_convst at AD7091R8 device init.

Signed-off-by: Marcelo Schmitt <marcelo.schmitt@analog.com>